Refactor scroll handling into `useScrollHandler` hook
Moved scroll-related logic from `HomeLayout` component to a dedicated custom hook `useScrollHandler`. This enhances code modularity and readability, making the component less cluttered and easier to maintain.
