+++ frontend-next-migration/src/app/[lng]/(home)/layout.tsx
@@ -1,10 +1,12 @@
-'use client';
-import { ReactNode, useEffect, useRef, useState } from 'react';
+'use client'
+import { ReactNode, useRef } from 'react';
 import { Navbar } from '@/widgets/Navbar';
 import { Footer } from '@/widgets/Footer';
 import { ScrollTop } from '@/features/ScrollTop';
 import Intro from './Intro';
-import {classNames} from "@/shared/lib/classNames/classNames";
+import { classNames } from "@/shared/lib/classNames/classNames";
+import { useScrollHandler } from './useScrollHandler';
+
 
 type Props = {
     children: ReactNode;
@@ -12,47 +14,7 @@ type Props = {
 
 export default function HomeLayout({ children }: Props) {
     const introRef = useRef<HTMLDivElement>(null);
-    const [isScrollbarHidden, setIsScrollbarHidden] = useState(true);
-
-    const scrollToContent = () => {
-        if (introRef.current) {
-            window.scrollTo({
-                top: introRef.current.clientHeight + 1,
-                behavior: 'smooth',
-            });
-        }
-    };
-    const updateScrollbarVisibility = () => {
-        const isBelowIntro = window.scrollY > (introRef.current?.clientHeight || window.innerHeight);
-        setIsScrollbarHidden(!isBelowIntro);
-    };
-    const scrollToIntro = (lastScrollY: { value: number }) => {
-        const currentScrollY = window.scrollY;
-        const scrollDirection = currentScrollY > lastScrollY.value ? 'down' : 'up';
-        lastScrollY.value = currentScrollY;
-        if (introRef.current) {
-            const introBottom = introRef.current.clientHeight;
-            if (scrollDirection === 'up' && window.scrollY <= introBottom) {
-                setTimeout(() => {
-                    window.scrollTo({
-                        top: 0,
-                        behavior: 'smooth',
-                    });
-                }, 100);
-            }
-        }
-    };
-    useEffect(() => {
-        let lastScrollY = {value : 0};
-        const handleScroll = () => {
-            updateScrollbarVisibility();
-            scrollToIntro(lastScrollY);
-        };
-        window.addEventListener('scroll', handleScroll, { passive: true });
-        return () => {
-            window.removeEventListener('scroll', handleScroll);
-        };
-    }, []);
+    const { isScrollbarHidden, scrollToContent} = useScrollHandler(introRef);
 
     return (
         <>
@@ -61,7 +23,7 @@ export default function HomeLayout({ children }: Props) {
             </div>
 
             <div className={classNames("main-content")}>
-                <Navbar/>
+                <Navbar />
                 {children}
                 <Footer />
                 <ScrollTop />
@@ -81,6 +43,3 @@ export default function HomeLayout({ children }: Props) {
         </>
     );
 }
-
-
-
